requirements.txt
beautifulsoup4==4.12.3
google-generativeai==0.3.2
matplotlib==3.8.2
openai==1.12.0
pandas==2.2.0
PyPDF2==3.0.1
requests==2.31.0
streamlit==1.31.1
stripe==7.12.0 